backend/requirements.txt
aniso8601==6.0.0
Click==7.1.2
Flask==2.0.1
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
python-dateutil==2.6.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
requests==2.31.0
python-dotenv==1.0.1
